fix: コンポーネントをトップ画面とランキング画面で分離
